fixed within sib models
Within sibling models needed to cal trait residuals to use for calculating relative to the family mean, so added that in to step 3. Trait residuals to be used in the between analysis are calculated in step 4 (had to do in different stages bc the way the code was already set up).